+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## x.x.x - 2021-10-xx
 
 * Changed: Delete `db-error.php` on uninstall or deactivate.
+* Added: Remove meta generator tags.
 
 ## 1.5.2 - 2021-10-03