+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## 1.5.3 - 2021-10-xx
+## 1.5.3 - 2021-10-05
 
 * Changed: Delete `db-error.php` on uninstall or deactivate.
 * Added: Remove meta generator tags.